[Refactor] 프로젝트를 멀티모듈 구조로 변환
BREAKING CHANGE: 기존 gradle 운영 방식과 다름!
