+++ app/src/main/java/fc/be/TripVoteApplication.java
@@ -1,0 +1,17 @@
+package fc.be;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication(scanBasePackageClasses = {
+        OpenAPIRoot.class,
+        NotificationRoot.class,
+        TripVoteApplication.class
+})
+public class TripVoteApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(TripVoteApplication.class, args);
+    }
+
+}
