+++ app/src/main/java/fc/be/domain/place/entity/Restaurant.java
@@ -1,0 +1,17 @@
+package fc.be.domain.place.entity;
+
+import fc.be.domain.place.Place;
+import jakarta.persistence.DiscriminatorValue;
+import jakarta.persistence.Entity;
+import lombok.AccessLevel;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import org.hibernate.annotations.Comment;
+
+@Entity
+@DiscriminatorValue("Restaurant")
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+@Getter
+@Comment("음식점")
+public class Restaurant extends Place {
+}
