+++ app/src/main/java/fc/be/exception/ApplicationException.java
@@ -1,0 +1,11 @@
+package fc.be.exception;
+
+import org.springframework.http.HttpStatusCode;
+import org.springframework.web.ErrorResponseException;
+
+public class ApplicationException extends ErrorResponseException {
+
+    public ApplicationException(HttpStatusCode status) {
+        super(status);
+    }
+}
