+++ app/src/main/resources/application.yml
@@ -1,0 +1,5 @@
+spring:
+  profiles:
+    include:
+      common
+    active: ${APP_PROFILE}
