+++ notification/build.gradle
@@ -1,0 +1,5 @@
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+}
