+++ notification/src/main/java/fc/be/NotificationRoot.java
@@ -1,0 +1,7 @@
+package fc.be;
+
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan(basePackageClasses = {NotificationRoot.class})
+public interface NotificationRoot {
+}
