+++ openapi/build.gradle
@@ -1,0 +1,10 @@
+bootJar { enabled = false }
+jar { enabled = true }
+
+dependencies {
+    // spring web
+    implementation 'org.springframework.boot:spring-boot-starter-web'
+
+    // xml jaxb
+    implementation 'jakarta.xml.bind:jakarta.xml.bind-api:+'
+}
