+++ openapi/src/main/java/fc/be/OpenAPIRoot.java
@@ -1,0 +1,7 @@
+package fc.be;
+
+import org.springframework.context.annotation.ComponentScan;
+
+@ComponentScan(basePackageClasses = {OpenAPIRoot.class})
+public interface OpenAPIRoot {
+}
