--- src/main/java/fc/be/tripvote/TripVoteApplication.java
@@ -1,13 +1,0 @@
-package fc.be.tripvote;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TripVoteApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TripVoteApplication.class, args);
-    }
-
-}
