--- src/main/java/fc/be/tripvote/domain/space/vo/VoteStatus.java
@@ -1,13 +1,0 @@
-package fc.be.tripvote.domain.space.vo;
-
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public enum VoteStatus {
-    VOTING("투표중"),
-    DONE("투표완료");
-
-    private final String description;
-}
