--- src/main/java/fc/be/tripvote/global/util/ContentTypeIdConverter.java
@@ -1,30 +1,0 @@
-package fc.be.tripvote.global.util;
-
-import fc.be.tripvote.domain.place.ContentTypeId;
-import jakarta.persistence.AttributeConverter;
-import jakarta.persistence.Converter;
-
-@Converter(autoApply = true)
-public class ContentTypeIdConverter implements AttributeConverter<ContentTypeId, Integer> {
-
-    @Override
-    public Integer convertToDatabaseColumn(ContentTypeId attribute) {
-        if (attribute == null) {
-            return null;
-        }
-        return attribute.getId();
-    }
-
-    @Override
-    public ContentTypeId convertToEntityAttribute(Integer dbData) {
-        if (dbData == null) {
-            return null;
-        }
-        for (ContentTypeId contentTypeId : ContentTypeId.values()) {
-            if (contentTypeId.getId() == dbData) {
-                return contentTypeId;
-            }
-        }
-        throw new IllegalArgumentException("Unknown database value: " + dbData);
-    }
-}
