--- src/main/resources/application-dev.yml
@@ -1,23 +1,0 @@
-spring:
-  config:
-    activate:
-      on-profile: dev
-
-  devtools:
-    livereload:
-      enabled: true
-
-  datasource:
-    driver-class-name: com.mysql.cj.jdbc.Driver
-    url: jdbc:mysql://localhost:3306/tripvote?characterEncoding=UTF-8
-    username: root
-    password: root
-
-  jpa:
-    show-sql: true
-    hibernate:
-      ddl-auto: create
-    properties:
-      hibernate:
-        format_sql: true
-        dialect: org.hibernate.dialect.MySQLDialect
