--- src/main/resources/application.yml
@@ -1,5 +1,0 @@
-spring:
-  profiles:
-    include:
-      core
-    active: ${APP_PROFILE}
