--- src/test/java/fc/be/tripvote/TripVoteApplicationTests.java
@@ -1,13 +1,0 @@
-package fc.be.tripvote;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TripVoteApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
